core/syslog: Add support for dynamic syslog identity strings

+++ src/misc.h
@@ -22,6 +22,10 @@
 
 
 
+#include <stddef.h>
+
+
+
 /*
  * Init/shutdown functions
  */
@@ -33,7 +37,7 @@ void snoopy_cleanup ();
 /*
  * String functions
  */
-void snoopy_string_append     (char *destString, const char *appendThis, int destStringMaxLength);
+void snoopy_string_append     (char *destString, const char *appendThis, size_t destStringMaxLength);
 int  snoopy_string_countChars (const char *stringToSearch, char characterToCount);
